fix clang compilation

+++ libft/libft_structsb/src/ft_sb_tostring.c
@@ -6,7 +6,7 @@
 /*   By: pepe <pepe@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/10 11:01:08 by jpriou            #+#    #+#             */
-/*   Updated: 2019/08/17 12:08:30 by pepe             ###   ########.fr       */
+/*   Updated: 2019/08/17 12:26:11 by pepe             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,22 +16,22 @@
 static void					calculate_length(void *elem, void *param)
 {
 	char					*str;
-	int						*size;
+	size_t					*size;
 
 	str = (char *)elem;
-	size = (int *)param;
+	size = (size_t *)param;
 	*size = *size + ft_strlen(str);
 }
 
 static void					build_str(void *elem, void *param1, void *param2)
 {
 	char					*str;
-	int						*actu;
+	size_t					*actu;
 	char					**res;
 	size_t					len;
 
 	str = (char *)elem;
-	actu = (int *)param1;
+	actu = (size_t *)param1;
 	res = (char **)param2;
 	len = ft_strlen(str);
 	ft_strncpy(*res + *actu - len, str, len);
@@ -41,7 +41,7 @@ static void					build_str(void *elem, void *param1, void *param2)
 char						*ft_sb_tostring(t_sb *sb)
 {
 	char	*res;
-	int		len_tot_stored;
+	size_t	len_tot_stored;
 
 	len_tot_stored = 0;
 	ft_slist_iterparam(sb, &len_tot_stored, calculate_length);
